chore: [#69] Bookmark ShareInfo Service Code Refacotr
